# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lotus RFQ Service API
  version: 0.1.0
  description: |
    Static API contract for the Lotus RFQ service.

    This spec separates implemented callable APIs from planned or unregistered APIs.
    Do not call endpoints marked `x-lotus-callable: false`.

    Execution safety notes:
    - RFQ accept may trigger execution only if the lane is operator-approved.
    - Matcher-ready lanes are not executable authority.
    - Execution may fail closed.
    - Settlement verification may delay final completion.
    - Ghost-fill protection can change final state.
    - Accounting updates only after settlement/finality.
    - Execution-scope token is not authority by itself.

    Funding safety notes:
    - Funding v0 is non-custodial funding preparation: Lotus returns route instructions and tracks user-broadcast transactions; users sign with their own wallets.
    - Lotus does not custody, sign, broadcast user wallet transactions, pool funds, or internally allocate user balances in v0.
    - LI.FI route quote does not equal ready-to-trade.
    - Execution preflight requires READY_TO_TRADE.
    - Funding Capability Matrix decides target chain/token/destination for capital movement.
    - Market Capability Matrix remains separate and decides market lane routeability/operator approval.
    - Withdrawal v0 is also non-custodial: Lotus records withdrawal intents, route previews, and user-broadcast tx hashes only.
...
    - Predict.fun is not PredictIt.
servers:
...
paths:
...
  /funding/intents:
    post:
      tags: [Funding]
      summary: Create a split-capable funding intent.
      description: Creates a funding intent only. It does not move funds, sign wallet transactions, or mark funds ready to trade.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: funding-intent
      x-lotus-danger: medium
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFundingIntentRequest"
      responses:
        "201": { $ref: "#/components/responses/FundingIntent" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /funding/intents/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Funding]
      summary: Read an owned funding intent.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: medium
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/FundingIntent" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /funding/intents/{id}/quote:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Funding]
      summary: Quote funding route legs through the Lotus LI.FI wrapper.
      description: Fails closed when LI.FI quotes are disabled or the Funding Capability Matrix cannot validate the destination. LI.FI is a route provider only; Lotus remains responsible for funding orchestration and readiness validation.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: funding-intent
      x-lotus-danger: medium
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/FundingIntent" }
        "409": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /funding/intents/{id}/submit:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Funding]
      summary: Record a user-broadcast funding transaction hash.
      description: Backend does not sign or broadcast transactions in v0. This endpoint records a tx hash after the user signs/broadcasts from their wallet.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: funding-intent
      x-lotus-danger: high
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFundingRouteLegRequest"
      responses:
        "202": { $ref: "#/components/responses/FundingIntent" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /funding/intents/{id}/status:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Funding]
      summary: Read frontend-safe aggregate and per-leg funding status.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/FundingIntent" }
        "404": { $ref: "#/components/responses/Error" }
  /funding/venue-balances:
    get:
      tags: [Funding]
      summary: Read frontend-safe venue-ready balances for the authenticated user.
      description: Returns a derived capital view from persisted READY_TO_TRADE funding records minus active withdrawal amounts. It does not imply Lotus custody or pooled balances.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/VenueBalances" }
        "401": { $ref: "#/components/responses/Error" }
  /funding/venue-activations:
    get:
      tags: [Funding]
      summary: Read frontend-safe venue balance activation requirements.
      description: Returns safe metadata for venue collateral activation. Signable approval transactions are returned only when backend operator-approved token/spender config exists; the frontend must never provide spender addresses.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: medium
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/VenueActivations" }
        "401": { $ref: "#/components/responses/Error" }
  /funding/history:
    get:
      tags: [Funding]
      summary: Read realtime-friendly funding and withdrawal history for the authenticated user.
      description: Returns sanitized funding and withdrawal timeline rows derived from persisted intents, route legs, latest reconciliation evidence, and tx hashes. It does not expose provider auth, raw venue payloads, private keys, or signer material.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 5
        - name: limit
          in: query
          required: false
          description: Backward-compatible alias for pageSize.
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200": { $ref: "#/components/responses/FundingHistory" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /funding/intents/{id}/receipt:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Funding]
      summary: Read sanitized funding receipt for an authenticated user's funding intent.
      description: Returns intent, leg, tx hash, reconciliation, readiness, and timestamp fields safe for user display. Raw provider payloads and secrets are excluded.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/GenericObject" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /funding/withdrawals/{id}/receipt:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Funding]
      summary: Read sanitized withdrawal receipt for an authenticated user's withdrawal intent.
      description: Returns withdrawal intent, source, leg, tx hash, reconciliation, completion, and timestamp fields safe for user display. Raw provider payloads and secrets are excluded.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/GenericObject" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /funding/withdrawals:
    post:
      tags: [Funding]
      summary: Create a split-capable withdrawal intent.
      description: Creates a non-custodial withdrawal intent from venue-ready capital. Lotus does not custody, sign, broadcast, call live venue withdrawal APIs, or mutate funding readiness records.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: funding-withdrawal-intent
      x-lotus-danger: medium
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWithdrawalIntentRequest"
      responses:
        "201": { $ref: "#/components/responses/WithdrawalIntent" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /funding/withdrawals/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Funding]
      summary: Read an owned withdrawal intent.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: medium
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/WithdrawalIntent" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /funding/withdrawals/{id}/quote:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Funding]
      summary: Quote non-custodial withdrawal route legs.
      description: Creates a frontend-safe withdrawal preview using Lotus withdrawal v0 scaffolding. It fails closed when venue withdrawal capability is disabled, source venue-ready balance is insufficient, destination wallet is invalid, capability state is unknown, or the venue is classified as AUTO_RESOLUTION_ONLY without an approved bridge-back path. Limitless beta support is limited to an explicitly enabled user-signed Base USDC to Solana USDC bridge-back route; its partner-managed backend withdrawal path is not wired to this endpoint.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: funding-withdrawal-intent
      x-lotus-danger: medium
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/WithdrawalIntent" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /funding/withdrawals/{id}/submit:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Funding]
      summary: Record a user-broadcast withdrawal transaction hash.
      description: Backend records the tx hash after the user signs/broadcasts. This does not submit a transaction, call a venue withdrawal API, or move funds from Lotus custody.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: funding-withdrawal-intent
      x-lotus-danger: high
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitWithdrawalRouteLegRequest"
      responses:
        "202": { $ref: "#/components/responses/WithdrawalIntent" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /funding/withdrawals/{id}/status:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Funding]
      summary: Read frontend-safe aggregate and per-leg withdrawal status.
      description: Status read is non-custodial and does not run live venue withdrawal mutation.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/WithdrawalIntent" }
        "404": { $ref: "#/components/responses/Error" }
  /funding/venues/capabilities:
    get:
      tags: [Funding]
      summary: Read frontend-safe venue funding capability matrix.
      description: Response redacts deposit addresses and provider secrets; it only exposes whether required server-side config exists. This Funding Capability Matrix is separate from market lane capability/operator approval.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Venue capability matrix.
          content:
            application/json:
              schema:
                type: object
                required: [capabilities]
                properties:
                  capabilities:
                    type: array
                    items: { $ref: "#/components/schemas/VenueCapability" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    FundingIntent:
      description: Frontend-safe funding intent response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FundingIntentResponse"
    VenueBalances:
      description: Frontend-safe derived venue-ready balance rows.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/VenueBalanceListResponse"
    VenueActivations:
      description: Frontend-safe venue balance activation requirements.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/VenueActivationListResponse"
...
  schemas:
...
    CreateFundingIntentRequest:
      type: object
      required: [sourceChain, sourceToken, sourceAmount, idempotencyKey, targets]
      additionalProperties: false
      properties:
        sourceChain: { type: string, example: SOLANA }
        sourceToken: { type: string, example: USDC }
        sourceAmount: { type: string, example: "100" }
        sourceWalletAddress:
          type: string
          description: External source wallet address. Optional when sourceWalletId is supplied or a default Solana Turnkey funding wallet exists.
        sourceWalletId:
          type: string
          description: Stored user wallet id. If supplied, backend resolves the address server-side and rejects wallets not owned by the user.
        idempotencyKey: { type: string }
        targets:
          type: array
          minItems: 1
          items:
            type: object
            required: [targetVenue]
            additionalProperties: false
            properties:
              targetVenue: { $ref: "#/components/schemas/Venue" }
              targetAmount: { type: string }
              targetPercentage: { type: number, minimum: 0, maximum: 100 }
    SubmitFundingRouteLegRequest:
      type: object
      required: [routeLegId, txHash]
      additionalProperties: false
      properties:
        routeLegId: { type: string }
        txHash: { type: string }
        status: { type: string }
...